feat: Connect PostList to Redux store
Connects the PostList component to the Redux store and creates the "posts" reducer.
Also creates the Root component, the wrapper component that provides us an easy way to set the initial state of the Redux store and thus allowing the creation of unit tests for components that is connected to Redux.