feat: PostList: Loading state and auto refresh
Adds a loading state to the PostList component. The state changes when the posts from the Redux store is populated.
Also was added an auto refresh, which cleans and fetches once again all posts. The idea behind this is to always show updated posts when the user visits the PostList component, but avoiding a refresh on every component mount.