fix: Refactor API calls to fetch correct range of data
Refactors the posts and comments API calls in order to fetch both latest available content, and also previous content (needed when the user clicks on the "load more" button).
A loading spinner was added when the users clicks on the "load more" buttons, so that the user know when it is actually doing some processing.